moving onto L2

+++ modules/quicklook/src/alg.py
@@ -196,57 +196,61 @@ class QuicklookAlg:
 
 
         #now onto the plotting of CCF
-        ccf_file = '/data/L2/20220524/KP.20220524.02360.58_L2.fits'
-        hdulist = fits.open(ccf_file)
-        print(hdulist.info())
-
-        ccf_color = ['GREEN_CCF','RED_CCF']
-        color_grid = ['Green','Red']
-
-        plt.rcParams.update({'font.size': 8})
-        plt.rcParams['legend.fontsize'] = plt.rcParams['font.size']
-        fig, ax = plt.subplots(1,1, sharex=True,figsize=(5,4))
-        ax = plt.subplot()
-        plt.subplots_adjust(left=0.15, bottom=0.15, right=0.95, top=0.9)
-        for i_color in range(len(ccf_color)):
-            ccf = np.array(hdulist[ccf_color[i_color]].data,'d')
-            ccf = np.sum(ccf,axis =0)#sum over orderlets
-            #print(np.shape(ccf))
-
-            step = float(self.config['RV']['step'])
-            vel_grid = np.array(range(-int(np.shape(ccf)[1]/2),int(np.shape(ccf)[1]/2),1),'d')*step
-            #print('step',step,len(vel_grid))
+        date = exposure_name[3:11]
+        ccf_file = '/data/L2/'+date+'/'+exposure_name+'_L2.fits'
+        print(date,ccf_file)
+        if os.path.exists(ccf_file):
+            print('Working on L2 file')
+            hdulist = fits.open(ccf_file)
 
 
-            mean_ccf = np.nanmean(ccf,axis = 0)/np.percentile(np.nanmean(ccf,axis = 0),[99.9])
-            print('test',np.shape(np.nanmean(ccf,axis = 0)))
+            ccf_color = ['GREEN_CCF','RED_CCF']
+            color_grid = ['Green','Red']
 
-            #mean_ccf = np.nanmedian(mean_ccf,axis = 0)
-            plt.plot(vel_grid,mean_ccf,label = ccf_color[i_color],color = color_grid[i_color],linewidth = 0.5)
+            plt.rcParams.update({'font.size': 8})
+            plt.rcParams['legend.fontsize'] = plt.rcParams['font.size']
+            fig, ax = plt.subplots(1,1, sharex=True,figsize=(5,4))
+            ax = plt.subplot()
+            plt.subplots_adjust(left=0.15, bottom=0.15, right=0.95, top=0.9)
+            for i_color in range(len(ccf_color)):
+                ccf = np.array(hdulist[ccf_color[i_color]].data,'d')
+                ccf = np.sum(ccf,axis =0)#sum over orderlets
+                #print(np.shape(ccf))
+
+                step = float(self.config['RV']['step'])
+                vel_grid = np.array(range(-int(np.shape(ccf)[1]/2),int(np.shape(ccf)[1]/2),1),'d')*step
+                #print('step',step,len(vel_grid))
+
+
+                mean_ccf = np.nanmean(ccf,axis = 0)/np.percentile(np.nanmean(ccf,axis = 0),[99.9])
+                print('test',np.shape(np.nanmean(ccf,axis = 0)))
+
+                #mean_ccf = np.nanmedian(mean_ccf,axis = 0)
+                plt.plot(vel_grid,mean_ccf,label = ccf_color[i_color],color = color_grid[i_color],linewidth = 0.5)
 
 
 
-            #fit the center of the ccf
-            fitter = modeling.fitting.LevMarLSQFitter()#the gaussian fit of the ccf
-            model = modeling.models.Gaussian1D()
-            fitted_model = fitter(model, vel_grid, 1.-mean_ccf)
-            gamma =fitted_model.mean.value
-            std =fitted_model.stddev.value
+                #fit the center of the ccf
+                fitter = modeling.fitting.LevMarLSQFitter()#the gaussian fit of the ccf
+                model = modeling.models.Gaussian1D()
+                fitted_model = fitter(model, vel_grid, 1.-mean_ccf)
+                gamma =fitted_model.mean.value
+                std =fitted_model.stddev.value
 
-            print(i_color,gamma,std)
+                print(i_color,gamma,std)
 
-            plt.plot([gamma,gamma],[np.nanmin(mean_ccf),1.],':',color ='gray',linewidth = 0.5)
-            ax.text(0.6,0.3+i_color*0.2,ccf_color[i_color]+' $\gamma$ (km/s): %5.2f' % gamma,transform=ax.transAxes)
-            ax.text(0.6,0.2+i_color*0.2,ccf_color[i_color]+' $\sigma$ (km/s): %5.2f' % std,transform=ax.transAxes)
+                plt.plot([gamma,gamma],[np.nanmin(mean_ccf),1.],':',color ='gray',linewidth = 0.5)
+                ax.text(0.6,0.3+i_color*0.2,ccf_color[i_color]+' $\gamma$ (km/s): %5.2f' % gamma,transform=ax.transAxes)
+                ax.text(0.6,0.2+i_color*0.2,ccf_color[i_color]+' $\sigma$ (km/s): %5.2f' % std,transform=ax.transAxes)
 
-        plt.xlabel('RV (km/s)')
-        plt.ylabel('CCF')
-        plt.title('Mean CCF')
-        plt.legend()
-        plt.savefig(output_dir+'fig/'+exposure_name+'_simple_ccf.pdf')
-        plt.savefig(output_dir+'fig/'+exposure_name+'_simple_ccf.png')
-        plt.close()
-
+            plt.xlabel('RV (km/s)')
+            plt.ylabel('CCF')
+            plt.title('Mean CCF')
+            plt.legend()
+            plt.savefig(output_dir+'fig/'+exposure_name+'_simple_ccf.pdf')
+            plt.savefig(output_dir+'fig/'+exposure_name+'_simple_ccf.png')
+            plt.close()
+        else: print('L2 file does not exist')
         #output the results to html
         f = open(output_dir+exposure_name+'_summary.html','w')
